requirements.txt
Flask==2.0.1
opencv-python-headless==4.5.3.56
numpy==1.21.2
Pillow==8.3.2
svglib==1.1.0
reportlab==3.6.1